# node-project/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "이 문서는 채용 공고 관리, 회원 인증, 지원 현황, 북마크, 기업 정보 조회 등을 제공하는 API에 대한 Swagger 문서입니다. 각 엔드포인트 별로 필요한 파라미터와 응답 형식을 한국어로 설명합니다."
  },
  "host": "113.198.66.75:443",
  "basePath": "/v1",
  "schemes": [
    "http"
  ],
  "paths": {
    "/jobs": {
      "get": {
        "description": "채용 공고 목록을 조회합니다...",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string",
            "description": "페이지 번호 (기본값: 1)"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string",
            "description": "페이지 당 표시할 공고 수 (기본값: 20)"
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string",
            "description": "정렬 기준 필드명 (기본값: createdAt)"
          },
          {
            "name": "order",
            "in": "query",
            "type": "string",
            "description": "정렬 순서 (asc 또는 desc, 기본값: desc)"
          },
          {
            "name": "지역",
            "in": "query",
            "type": "string",
            "description": "채용 공고의 지역 필터"
          },
          {
            "name": "경력",
            "in": "query",
            "type": "string",
            "description": "경력 조건 필터 (예: 신입, 경력)"
          },
          {
            "name": "minSalary",
            "in": "query",
            "type": "string",
            "description": "최소 연봉 필터"
          },
          {
            "name": "maxSalary",
            "in": "query",
            "type": "string",
            "description": "최대 연봉 필터"
          },
          {
            "name": "직무분야",
            "in": "query",
            "type": "string",
            "description": "직무 분야 필터 (쉼표로 구분 가능)"
          },
          {
            "name": "keyword",
            "in": "query",
            "type": "string",
            "description": "키워드 검색 (회사명, 직무분야 등)"
          },
          {
            "name": "회사명",
            "in": "query",
            "type": "string",
            "description": "회사명 검색"
          },
          {
            "name": "제목",
            "in": "query",
            "type": "string",
            "description": "공고 제목 검색"
          }
        ],
        "responses": {
          "200": {
            "description": "공고 조회 성공"
          },
          "500": {
            "description": "서버 오류 발생"
          }
        }
      },
      "post": {
        "description": "새로운 채용 공고를 추가합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "추가할 공고 정보",
            "schema": {
              "type": "object",
              "properties": {
                "제목": {
                  "example": "신입 백엔드 개발자 채용",
                  "description": "공고 제목",
                  "required":true
                },
                "회사명": {
                  "example": "테크스타트업",
                  "description": "회사명",
                  "required":true
                },
                "지역": {
                  "example": "서울",
                  "description": "지역"
                },
                "경력": {
                  "example": "신입",
                  "description": "경력 요구사항"
                },
                "직무분야": {
                  "example": "백엔드",
                  "description": "직무 분야"
                },
                "연봉정보": {
                  "example": "면접 후 협의",
                  "description": "연봉 정보"
                },
                "조회수": {
                  "example": "0",
                  "description": "초기 조회수 (옵션)"
                },
                "마감일": {
                  "example": "2024-11-24",
                  "description": "공고 마감일",
                  "required":true
                },
                "링크": {
                  "example": "https://www.google.com",
                  "description": "링크 입력",
                  "required":true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "공고 생성 성공"
          },
          "500": {
            "description": "서버 오류 발생"
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "description": "특정 ID의 채용 공고 상세 정보를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "조회할 공고의 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "공고 상세 조회 성공"
          },
          "404": {
            "description": "공고를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류 발생"
          }
        }
      },
      "put": {
        "description": "특정 ID의 채용 공고 정보를 수정합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "수정할 공고의 ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "수정할 공고 정보",
            "schema": {
              "type": "object",
              "properties": {
                "제목": {
                  "example": "경력 백엔드 개발자 채용",
                  "description": "수정할 공고 제목(옵션)"
                },
                "회사명": {
                  "example": "테크스타트업",
                  "description": "회사명(옵션)"
                },
                "지역": {
                  "example": "서울",
                  "description": "지역(옵션)"
                },
                "경력": {
                  "example": "3년 이상",
                  "description": "경력 요구사항(옵션)"
                },
                "직무분야": {
                  "example": "백엔드",
                  "description": "직무 분야(옵션)"
                },
                "연봉정보": {
                  "example": "연봉 4000~5000만원",
                  "description": "연봉 정보(옵션)"
                },
                "조회수": {
                  "example": "100",
                  "description": "조회수(옵션)"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "공고 수정 성공"
          },
          "404": {
            "description": "공고를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류 발생"
          }
        }
      },
      "delete": {
        "description": "특정 ID의 공고를 삭제합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "삭제할 공고의 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "공고 삭제 성공"
          },
          "404": {
            "description": "공고를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류 발생"
          }
        }
      }
    },
...
  }
}